v0.75-2

+++ install.sh
@@ -12,6 +12,8 @@ install -m 755 blconf /usr/sbin
 #install -m 755 geturl /usr/sbin
 install -m 644 dosdisk.img /usr/share/biosdisk
 install -m 644 biosdisk.conf /etc/
-install -m 644 biosdisk-mkrpm-redhat-template.spec /usr/share/biosdisk
-install -m 644 biosdisk-mkrpm-generic-template.spec /usr/share/biosdisk
+if [ -e /usr/bin/rpm ] || [ -e /bin/rpm ]; then
+    install -m 644 biosdisk-mkrpm-redhat-template.spec /usr/share/biosdisk
+    install -m 644 biosdisk-mkrpm-generic-template.spec /usr/share/biosdisk
+fi
 install -m 644 biosdisk.8.gz /usr/share/man/man8
